+++ slushy-app/src/main/java/net/kemitix/slushy/app/DueDays.java
@@ -1,0 +1,8 @@
+package net.kemitix.slushy.app;
+
+public interface DueDays {
+    /**
+     * How long before a still-waiting-to-be-read response should be sent.
+     */
+    long getDueDays();
+}
