--- slushy-app/src/main/java/net/kemitix/slushy/app/duedates/DueIn30DaysRoute.java
@@ -1,27 +1,0 @@
-package net.kemitix.slushy.app.duedates;
-
-import org.apache.camel.builder.RouteBuilder;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-
-@ApplicationScoped
-public class DueIn30DaysRoute
-        extends RouteBuilder {
-    private final SetDueInDays setDueInDays;
-
-    @Inject
-    public DueIn30DaysRoute(SetDueInDays setDueInDays) {
-        this.setDueInDays = setDueInDays;
-    }
-
-    @Override
-    public void configure() {
-        from("direct:Slushy.DueIn30Days")
-                .routeId("Slushy.DueIn30Days")
-                .setHeader("SlushyDueInDays")
-                .constant("30")
-                .bean(setDueInDays)
-        ;
-    }
-}
