+++ slushy-app/src/main/java/net/kemitix/slushy/app/inbox/InboxConfig.java
@@ -1,13 +1,9 @@
 package net.kemitix.slushy.app.inbox;
 
+import net.kemitix.slushy.app.DueDays;
 import net.kemitix.slushy.app.ListProcessConfig;
 
 public interface InboxConfig
-        extends ListProcessConfig {
-
-    /**
-     * How long before a still-waiting-to-be-read response should be sent.
-     */
-    long getDueDays();
+        extends ListProcessConfig, DueDays {
 
 }
