+++ slushy/src/main/resources/application.properties
@@ -24,7 +24,7 @@ slushy.inbox.routing-slip = \
   direct:Slushy.ReformatCard,\
   direct:Slushy.VerifyAttachmentValidType,\
   direct:Slushy.MultiSubMonitor,\
-  direct:Slushy.DueIn30Days,\
+  direct:Slushy.Inbox.ResponseDueBy,\
   direct:Slushy.AddMember,\
   direct:Slushy.Inbox.MoveToTargetList,\
   direct:Slushy.Inbox.SendEmailConfirmation
